# docs/openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: Lyra API
  version: 1.0.0
  description: |
    Client-agnostic HTTP API for the Lyra self-hosted mail client.

    - Product routes live under `/api/v1`.
    - Unversioned ops probes: `GET /health` (200, empty body), `GET /version` (JSON).
    - Errors use `{ "error": string, "code"?: string }`.
    - See `docs/specs/2026-08-26-lyra-http-api-surface.md` for web client boundary and v2 policy.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/settings/captcha:
    get:
      tags: [settings]
      summary: Login captcha settings
      description: |
        Server-wide captcha config for login/bootstrap. Each provider keeps
        its own site-key/secret pair so switching providers never discards
        credentials; `active` picks which pair protects login. The secret is
        write-only — responses only report `hasSecret`. `source` tells whether
        the config came from a saved setting or the `LYRA_CAPTCHA_*`
        env vars; a saved setting overrides the env config. When a setting is
        saved, the env pair is merged into `providers` for any provider the
        setting does not cover.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaSettings"
    put:
      tags: [settings]
      summary: Set login captcha provider and key pairs
      description: |
        `active: "none"` disables the captcha entirely (also overriding any
        env config). `providers` merges per-provider updates: an object
        upserts that pair (omitted fields keep the stored value), `null`
        removes the pair. The active provider must resolve to a complete pair
        (from the settings or the env config for the same provider).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active:
                  type: string
                  enum: [none, turnstile, hcaptcha, recaptcha, recaptcha-v3]
                providers:
                  type: object
                  additionalProperties:
                    oneOf:
                      - type: "null"
                      - type: object
                        properties:
                          siteKey:
                            type: string
                          secret:
                            type: string
                            description: Omit/empty to keep the stored secret.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaSettings"
        "400":
          description: Unknown provider, or the active provider lacks a site key/secret
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session
      description: |
        Token from `POST /api/v1/auth/login` or `/api/v1/auth/bootstrap`.
        Send as `Authorization: Bearer <token>` on protected routes.
        Invalid or expired sessions return 401 with `{ "error", "code": "unauthorized" }`.
...
  schemas:
...
    CaptchaSettings:
      type: object
      properties:
        active:
          type: string
          enum: [none, turnstile, hcaptcha, recaptcha, recaptcha-v3]
          description: The provider currently protecting login.
        providers:
          type: object
          description: Known key pairs keyed by provider (secret never included).
          additionalProperties:
            type: object
            properties:
              siteKey:
                type: string
              hasSecret:
                type: boolean
                description: The secret is write-only; this reports whether one is stored.
        source:
          type: string
          enum: [settings, env]
          description: Saved setting (overrides) or LYRA_CAPTCHA_* env vars.
...
security:
  - bearerAuth: []